+++ app/src/test/java/org/simple/clinic/selectcountry/SelectCountryUiRendererTest.kt
@@ -11,13 +11,13 @@ class SelectCountryUiRendererTest {
   private val defaultModel = SelectCountryModel.FETCHING
   private val ui = mock<SelectCountryUi>()
 
-  private val india = TestData.countryV2(
+  private val india = TestData.country(
       isoCountryCode = "IN",
       deploymentEndPoint = "https://in.simple.org",
       displayName = "India",
       isdCode = "91"
   )
-  private val bangladesh = TestData.countryV2(
+  private val bangladesh = TestData.country(
       isoCountryCode = "BD",
       deploymentEndPoint = "https://bd.simple.org",
       displayName = "Bangladesh",